feat(core): add module loader
