+++ packages/core/src/types.ts
@@ -1,0 +1,5 @@
+export interface ModuleMetadata {
+  name: string;
+  version: string;
+  description?: string;
+}
